requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.2
pandas==2.2.3
psutil==6.1.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
Werkzeug==3.1.3
